// src/types/Related_Info.types.ts
    export interface Related_Info_Form_Types {
        domicile: string;
        maritalStatus: string;
        currentAddress: string;
        country: string;
        placeOfBirth: string;
        religion: string;
        citizenIdentification: string;
        dateOfIssue: string;
        placeOfIssue: string;
        resident: string;
        bankAccountNumber: string;
        bankAccountName: string;
        bankName: string;
        personalTaxCode: string;
        epfNo: string;
        socialSecurityNo: string;
        facebook: string;
        linkedin: string;
        skype: string;
    }
   export type Input_Event_Type = React.ChangeEvent<HTMLInputElement>